+++ src/core/component/directives/render/interface.ts
@@ -6,6 +6,6 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import type { DirectiveBinding, VNode } from 'core/component/engines';
+import type { DirectiveBinding, VNode, SSRBufferItem } from 'core/component/engines';
 
-export interface DirectiveParams extends DirectiveBinding<CanUndef<CanArray<VNode>>> {}
+export interface DirectiveParams extends DirectiveBinding<CanUndef<CanArray<VNode> | SSRBufferItem>> {}